pyproject.toml
[project]
name = "COBRA"
version = "1.0.0"
description = "Charting Outlines by Recurrent Adaptation – A Deep Contour Detector developed for Charting Greenland's Glacier Calving Fronts."
authors = [
    { name = "Konrad Heidler", email = "k.heidler@tum.de" }
]
dependencies = [
    "numpy>=1.0.0",
    "wandb>=0.17.7",
    "pyyaml>=6.0.2",
    "jax[cuda12_pip]>=0.4.31",
    "optax>=0.2.3",
    "augmax>=0.3.4",
    "einops>=0.8.0",
    "matplotlib>=3.9.2",
    "scikit-image>=0.24.0",
    "tqdm>=4.0.0",
    "pillow>=9.0.0",
    "torch==2.4.0+cpu",
    "tensorflow-datasets>=4.9.6",
    "tensorflow[cpu]>=2.17.0",
    "flax>=0.10.1",
    "pyrallis>=0.3.1",
    "black>=24.10.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ipython>=8.26.0",
    "pyright>=1.1.388",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/zakynthos"]

[[tool.rye.sources]]
name = "jax"
# url = "https://jorenham.github.io/jax_pep503/"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
type = "find-links"  # this is implied

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
